requirements.txt
altair==5.2.0
annotated-types==0.6.0
attrs==23.1.0
blinker==1.7.0
blis>=1.3.0,<2.0.0
cachetools==5.3.2
catalogue==2.0.10
certifi==2023.11.17
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
cloudpathlib==0.16.0
confection==0.1.4
configobj==5.0.8
configparser==6.0.0
cymem==2.0.8
etelemetry==0.3.1
filelock==3.13.1
future==0.18.3
gitdb==4.0.11
GitPython==3.1.40
httplib2==0.22.0
idna==3.6
importlib-metadata==6.11.0
isodate==0.6.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
langcodes==3.3.0
looseversion==1.3.0
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
murmurhash==1.0.10
networkx==3.2.1
nibabel==5.1.0
nipype==1.8.6
nltk==3.8.1
numpy==1.26.2
packaging==23.2
pandas==2.1.3
pathlib==1.0.1
Pillow==10.1.0
preshed==3.0.9
protobuf==4.25.1
prov==2.0.0
pyarrow==14.0.1
pydantic>=2.0,<3.0
pydantic_core>=2.0,<3.0
pydeck==0.8.1b0
pydot==1.4.2
Pygments==2.17.2
PyMuPDF==1.23.7
PyMuPDFb==1.23.7
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
pyxnat==1.6
rdflib==7.0.0
referencing==0.31.1
regex==2023.10.3
requests==2.31.0
rich==13.7.0
rpds-py==0.13.2
scipy==1.11.4
simplejson==3.19.2
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
spacy>=3.8.0,<4.0.0
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
streamlit==1.29.0
tenacity==8.2.3
thinc>=8.3.0,<9.0.0
toml==0.10.2
toolz==0.12.0
tornado==6.4
tqdm==4.66.1
traits==6.3.2
typer==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.1.0
validators==0.22.0
wasabi==1.1.2
watchdog==3.0.0
weasel>=0.4.0,<1.0.0
zipp==3.17.0
python-dotenv==1.0.0
bcrypt==4.1.2
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
slowapi==0.1.9
